+++ catalyst_voices/packages/catalyst_voices_models/lib/src/account/account_role.dart
@@ -1,0 +1,5 @@
+enum AccountRole {
+  voter,
+  proposer,
+  drep,
+}
