+++ catalyst_voices/packages/catalyst_voices_models/lib/src/catalyst_voices_models.dart
@@ -1,5 +1,6 @@
 library catalyst_voices_models;
 
+export 'account/account_role.dart';
 export 'auth/authentication_status.dart';
 export 'auth/password_strength.dart';
 export 'errors/errors.dart';
